# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API",
        "title": "Swagger UI",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/account/balance": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Получить инфу о балансе",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chain.BalanceInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/chain.BalanceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/account/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Получение аккаунта по мнемонику",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chain.CreateAccountInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/chain.AccountResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/account/mnemonic": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Создание мнемоника",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chain.CreateMnemonicInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/account/restore": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Получение аккаунта по ключу",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chain.RestoreAccountInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/chain.AccountResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/chains/all": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chain"
                ],
                "summary": "Получение данных о сетях",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/chain.ShortResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/transaction/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Отправить транзакцию",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chain.SendTxInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/chain.SendTxResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/transaction/simulate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Симуляция транзакции для расчета параметров",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chain.SimulateTxInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/chain.SimulateTxResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "chain.AccountResponse": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "key": {
                    "type": "string"
                }
            }
        },
        "chain.BalanceInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
        "chain.BalanceResponse": {
            "type": "object",
            "properties": {
                "availableAmount": {
                    "type": "string"
                },
                "stakedAmount": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "string"
                }
            }
        },
        "chain.CreateAccountInput": {
            "type": "object",
            "properties": {
                "accountPath": {
                    "type": "integer"
                },
                "chainPrefixes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coinType": {
                    "type": "integer"
                },
                "indexPath": {
                    "type": "integer"
                },
                "mnemonic": {
                    "type": "string"
                }
            }
        },
        "chain.CreateMnemonicInput": {
            "type": "object",
            "properties": {
                "mnemonicSize": {
                    "type": "integer"
                }
            }
        },
        "chain.RestoreAccountInput": {
            "type": "object",
            "properties": {
                "chainPrefixes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "key": {
                    "type": "string"
                }
            }
        },
        "chain.SendTxInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "gasAdjusted": {
                    "type": "string"
                },
                "gasPrice": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "chain.SendTxResponse": {
            "type": "object",
            "properties": {
                "txHash": {
                    "type": "string"
                }
            }
        },
        "chain.ShortResponse": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string"
                },
                "bech32Prefix": {
                    "type": "string"
                },
                "chainId": {
                    "type": "string"
                },
                "chainName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display": {
                    "type": "string"
                },
                "keyAlgos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "logoPngUrl": {
                    "type": "string"
                },
                "logoSvgUrl": {
                    "type": "string"
                },
                "prettyName": {
                    "type": "string"
                },
                "slip44": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "chain.SimulateTxInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "chain.SimulateTxResponse": {
            "type": "object",
            "properties": {
                "averageGasPrice": {
                    "type": "string"
                },
                "gasAdjusted": {
                    "type": "string"
                },
                "highGasPrice": {
                    "type": "string"
                },
                "lowGasPrice": {
                    "type": "string"
                }
            }
        },
        "http.apiResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "isSuccess": {
                    "type": "boolean"
                },
                "result": {}
            }
        }
    }
}